setting up playbook infrastructure
This will have to move over to the pi shortly, because we can't easily
test this on a Mac.

We're going to make the assumption that these machines are only managed
by the Ansible scripts. We have to do a bunch of machine-specific
config, like detecting wifi adapters, so "bare metal" work is necessary.

It's old-skool systems administration...